+++ PCbuild/pythoncore.vcxproj
@@ -232,7 +232,6 @@
     <ClInclude Include="..\Include\internal\pycore_crossinterp.h" />
     <ClInclude Include="..\Include\internal\pycore_crossinterp_data_registry.h" />
     <ClInclude Include="..\Include\internal\pycore_debug_offsets.h" />
-    <ClInclude Include="..\Include\internal\pycore_remote_debug.h" />
     <ClInclude Include="..\Include\internal\pycore_descrobject.h" />
     <ClInclude Include="..\Include\internal\pycore_dict.h" />
     <ClInclude Include="..\Include\internal\pycore_dict_state.h" />
