+++ Python/remote_debugging.c
@@ -6,7 +6,7 @@
 #include "internal/pycore_ceval.h"
 
 #if defined(Py_REMOTE_DEBUG) && defined(Py_SUPPORTS_REMOTE_DEBUG)
-#include "internal/pycore_remote_debug.h"
+#include "remote_debug.h"
 
 static int
 init_proc_handle(proc_handle_t *handle, pid_t pid) {
